- Added scrolling menu for 40 homebrews max (thanks to MrShawman!) - Separate mallocs for p5 and p2 (r80 and r81 malloc code was wrong) - Included missing file lib.h - Included some code for loading utility modules (not operating yet) --> IMHO HBL needs a serious recoding effort...
git-svn-id: https://valentine-hbl.googlecode.com/svn/trunk@82 ae707b24-f6f8-a420-3b91-9542b9ad9fd8

+++ eLoader/config.c
@@ -1,7 +1,7 @@
+#include "sdk.h"
 #include "debug.h"
 #include "config.h"
 #include "utils.h"
-#include "lib.h"
 
 /* LOCAL GLOBALS :( */
 
@@ -43,7 +43,7 @@ int config_initialize()
     if (!file_exists(buffer))
         strcpy(buffer, IMPORTS_PATH);
    
-    LOGSTR1("Config file:%s\n", (ULONG) buffer);
+    LOGSTR1("Config file:%s\n", (u32) buffer);
     
 	g_config_file = sceIoOpen(buffer, PSP_O_RDONLY, 0777);
 	return g_config_file;
